Update 2015-08-20-recursive-types-and-folds.md
This PR changes "type" to "case" in [the section heading](https://fsharpforfunandprofit.com/posts/recursive-types-and-folds/#the-relationship-between-the-type-constructors-and-the-handlers) "The relationship between the type constructors and the handlers".

Marco Rodriguez asked in [this comment](https://fsharpforfunandprofit.com/posts/recursive-types-and-folds/#comment-5022605232) if this use of "type constructor" was a typo.  I think he is correct.